pyproject.toml
[tool.poetry]
name = "lms-backend"
version = "0.1.0"
description = ""
authors = ["mont3ll <meluseno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.0"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
redis = "^5.2.1"
celery = "^5.5.0"
requests = "^2.32.3"
dj-database-url = "^2.3.0"
dotenv = "^0.9.9"
django-storages = "^1.14.6"
boto3 = "^1.35.0"
django-filter = "^25.1"
drf-nested-routers = "^0.94.1"
pillow = "^11.2.1"
whitenoise = "^6.9.0"
reportlab = "^4.4.2"
anthropic = "^0.75.0"
# LTI 1.3 integration
pylti1p3 = "^2.0.0"
# SAML SSO integration
python3-saml = "^1.16.0"
cryptography = "^46.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
